requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2.1
django-import-export==4.3.7
djangoql==0.18.1
gunicorn==23.0.0
packaging==25.0
ply==3.11
sqlparse==0.5.3
tablib==3.8.0
psycopg2-binary==2.9.10
whitenoise
